% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwsem.R
\name{generalized.covariance}
\alias{generalized.covariance}
\title{Generalized covariance function}
\usage{
generalized.covariance(R1, R2)
}
\arguments{
\item{R1}{a numerical vector of residuals}

\item{R2}{a second numerical vector of residuals}
}
\value{
T.stat: The test statistic, which is asymptotically distributed
as a standard normal variate;

prob: asymptotic null probability of the T statistic.
}
\description{
This function calculates the generalized covariance statistic of
Shah, R.D. & Peters, J. (2020); i.e. Y1 \emph{|} Y2 |{C}, where C is a set of
common conditioning variables. Typically, R1 and R2 are the residuals
from pairs of regressions of two dependent variables (Y1 and Y2) on
a set of conditioning variables.

Shah, R.D. & Peters, J. (2020). The hardness of conditional independence
testing and the generalized covariance measure.  The Annals of Statistics
48:1514-1538.
}
\examples{
generalized.covariance(R1,R2)
}
