% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_data.R
\docType{data}
\name{nested_data}
\alias{nested_data}
\title{nested_data:
An example with nesting structure}
\format{
\subsection{'nested_data'}{

\describe{
\item{year}{Year observation is collected}
\item{nest}{Nest in which the observation is collected}
\item{ind}{the individual within a given nest}
\item{XR}{A binary variable}
\item{XM}{A Gaussian variable}
\item{XH}{A Gaussian variable}
\item{HP}{A Guassian variable}
\item{XF}{A Gaussian variable}
}
}
}
\usage{
nested_data
}
\description{
nested_data:
An example with nesting structure
}
\keyword{datasets}
