% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{sim_data:
Simulated data with correlated errors involving endogenous
variables and without any grouping structure.  Data generated
using this mixed acyclic graph:
X1->X2->X3->X4 and X2<->X4}
\format{
\subsection{'sim_data'}{

A data set with 100 rows and 4 columns
\describe{
\item{X1}{A Guassian variable}
\item{X2}{A Poisson variable}
\item{X3}{A Poisson variable}
\item{X4}{A Guassian variable}
}
}
}
\usage{
sim_data
}
\description{
sim_data:
Simulated data with correlated errors involving endogenous
variables and without any grouping structure.  Data generated
using this mixed acyclic graph:
X1->X2->X3->X4 and X2<->X4
}
\keyword{datasets}
