% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_normal.with.nesting.R
\docType{data}
\name{sim_normal.with.nesting}
\alias{sim_normal.with.nesting}
\title{sim_normal.with.nesting:
Simulated data with correlated errors involving endogenous
variables, normally-distributed data and without any grouping structure
Data generated using this mixed acyclic graph:
X1->X2->X3->X4 and X2<->X4}
\format{
\subsection{'sim_normal.with.nesting'}{

A data set with 100 rows and 5 columns. The last column "group" holds
the grouping structure
\describe{
\item{X1}{A Gaussian variable}
\item{X2}{A Gaussian variable}
\item{X3}{A Gaussian variable}
\item{X4}{A Gaussian variable}
\item{group}{A character variable holding the group names}
}
}
}
\usage{
sim_normal.with.nesting
}
\description{
sim_normal.with.nesting:
Simulated data with correlated errors involving endogenous
variables, normally-distributed data and without any grouping structure
Data generated using this mixed acyclic graph:
X1->X2->X3->X4 and X2<->X4
}
\keyword{datasets}
