% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tetrads.R
\docType{data}
\name{sim_tetrads}
\alias{sim_tetrads}
\title{sim_tetrads:
Simulated data to be used with the vanishing.tetrads function
Data generated using this directed acyclic graph, with L being latent:
L->X1, L->X2, L->X3->X4}
\format{
\subsection{'sim_tetrads'}{

A data set with 500 rows and 4 columns
\describe{
\item{X1}{A Gaussian variable}
\item{X2}{A Gaussian variable}
\item{X3}{A Gaussian variable}
\item{X4}{A Gaussian variable}
}
}
}
\usage{
sim_tetrads
}
\description{
sim_tetrads:
Simulated data to be used with the vanishing.tetrads function
Data generated using this directed acyclic graph, with L being latent:
L->X1, L->X2, L->X3->X4
}
\keyword{datasets}
