% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwsem.R
\name{summary.pwSEM.class}
\alias{summary.pwSEM.class}
\title{Summary Method for pwSEM Class}
\usage{
\method{summary}{pwSEM.class}(object, structural.equations = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{pwSEM}.}

\item{structural.equations}{Logical (TRUE outputs the structural equations)}

\item{...}{other arguments}
}
\value{
A summary of the pwSEM object.
}
\description{
Provides a summary for objects of class \code{pwSEM}.
}
