% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwsem.R
\name{update.fun}
\alias{update.fun}
\title{updates a gamm4 or gam model by adding the terms in "add.terms"}
\usage{
update.fun(sem.functions, i, all.grouping.vars, add.terms, data)
}
\arguments{
\item{sem.functions}{a list}

\item{i}{a single value indexing the function in sem.functions}

\item{all.grouping.vars}{a vector}

\item{add.terms}{a vector}

\item{data}{a data frame}
}
\description{
updates a gamm4 or gam model by adding the terms in "add.terms"
}
