% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwsem.R
\name{CI.algorithm}
\alias{CI.algorithm}
\title{The CI.algorithm function}
\usage{
CI.algorithm(
  dat,
  family = NA,
  nesting = NA,
  smooth = TRUE,
  alpha.reject = 0.05,
  write.result = T
)
}
\arguments{
\item{dat}{A data frame containing the variables for which a partially-
oriented dependency graph is sought.  All variables in this data frame
will be used except for those listed in the nesting= argument; these
variables describe the nesting structure (if present) of the other
variables.}

\item{family}{A  data frame giving the name of the distributional type
of each variable that is not Gaussian. Example: data.frame(X1="binomial")
if all variables except X1 are Gaussian but X1 is binomial.}

\item{nesting}{A named list.  Each name in the list is the name of the variable
in the data set (except for the nesting= variables) followed by a character
vector giving the names of the variables in dat holding the nesting structure.
The default is nesting=NA, which is used if there are no nesting variables.}

\item{smooth}{A logical value stating if linear (or generalized linear)
links between the variables are assumed (smooth=FALSE) or not (smooth=FALSE).
The default is smooth=TRUE, in which case generalized additive models are
used to measure conditional independence and therefore non-linear relationships
are assumed.}

\item{alpha.reject}{A numerical value between 0 and 1 giving the
"significance level" to use when judging (conditional) independence.  The
default value is alpha.reject=0.05.}

\item{write.result}{A logical value indicating if you want the resulting
partially-oriented dependency graph to be output to the screen
(write.result = T) or just the adjacency matrix returned as output.  The
default value is write.result = T.}
}
\value{
Just the partially-oriented graph output to the screen or
just the adjacency matrix
}
\description{
This function impliments the exploratory method of causal
discovery called the CI (Causal Inference) algorithm of Pearl.
}
\examples{
CI.algorithm(dat=nested_data,family=c("gaussian","gaussian","gaussian","gaussian"
),nesting=NA,smooth=TRUE,alpha.reject=0.05)

}
